+++ OntraportApi/IOntraportForms.cs
@@ -12,7 +12,7 @@ public interface IOntraportForms : IOntraportBaseRead<ApiForm>
     /// </summary>
     /// <param name="formId">The ID of the form to retrieve HTML for.</param>
     /// <returns>The form HTML.</returns>
-    Task<string?> SelectSmartFormHtmlAsync(int formId, CancellationToken cancellationToken = default);
+    Task<string?> SelectSmartFormHtmlAsync(long formId, CancellationToken cancellationToken = default);
 
     /// <summary>
     /// Retrieves name and ID pairs for all existing form blocks.
